+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added TLS fields to `elasticsearch` output.
+- New Bloblang methods `encrypt_aes` and `decrypt_aes` added.
 
 ## 3.21.0 - 2020-07-12